rename: BaseEntity entity 패키지로 이동 (#5)
