+++ src/main/java/team05/integrated_feed_backend/module/post/entity/Hashtag.java
@@ -2,12 +2,10 @@ package team05.integrated_feed_backend.module.post.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
-import team05.integrated_feed_backend.common.BaseEntity;
+import team05.integrated_feed_backend.common.entity.BaseEntity;
 
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Set;
 
 @Entity
 @Getter
