+++ src/main/java/team05/integrated_feed_backend/module/post/entity/PostHashtag.java
@@ -1,7 +1,7 @@
 package team05.integrated_feed_backend.module.post.entity;
 import jakarta.persistence.*;
 import lombok.*;
-import team05.integrated_feed_backend.common.BaseEntity;
+import team05.integrated_feed_backend.common.entity.BaseEntity;
 
 @Entity
 @Getter
